+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - renamed command `odoo-helper print_config` to `odoo-helper print-config`
 - Added `odoo-helper test --coverage-skip-covered` option
 - Added `odoo-helper addons update-py-deps` command
+- Added aliase `odoo-helper-log`
 - Removed old unused options
   - `odoo-helper --addons-dir <addons_directory>`
   - `odoo-helper --downloads-dir <downloads_directory>`